Updated documentation, fixed error in CtoS, added helper functions
Updated the documention found in the doc strings for all functions.

Found and corrected an error in how the order of time constants are returned in the C to S function.

Added two helper functions to preset the flags for finding the modulus at time of material functions. Updated the convolution funtions to use these helper functions, to be more readable.